requirements.txt
Django==2.2.6
psycopg2==2.8.4

pillow==6.2.1
djangorestframework==3.10.3
django-cors-headers==3.1.1
psycopg2-binary==2.8.4
pandas==0.25.3